Fixed the leaderboard
The leaderboard is now finished and working, even though its not the best code or even good code, I was quite stressed making it but it's done and works pretty decently, you could probably break it really easily. I fixed the restarting issues as well as tidying up the project abit.